easy_my_trip_holidays/flight_agent/requirements.txt
a2a-sdk
uvicorn
pydantic
python-dotenv
